fix(nai-122): in-scope-stretch — sign-extend negative param int32 values through paramLookup / InvTotalParam / hunt sumPlayerInvParam
Surfaced at NAI-122 smoke handoff: bronze weapons display crush/slash/
stab bonuses as 4294967292 / 4294967294 in the equipment UI (= int32
-4 / -2 read as uint32). Same root cause likely produces the
"all hits deal zero damage" residual #1 — adjacent equipment slots
have small negative bonuses (RuneScape weapon configs encode -2/-4
crush penalties on most stab/slash weapons), so OBJ params summed
into combat-stat varps overflowed via uint32→int direct conversion,
producing absurd values that downstream max-hit arithmetic likely
clamps or saturates.

DecodeParams (pkg/objtype/paramtype.go:21) reads param ints via
Packet.G4 (uint32 return) and stores them as the raw uint32 bit
pattern. Three consumer sites then converted via int(iv) which
zero-extends on 64-bit Go (int = int64), losing sign:

  pkg/script/handlers_config.go:38 (paramLookup — generic OC/NC/LC/STRUCT_PARAM)
  pkg/script/handlers_inv.go:249 (INV_TOTALPARAM)
  modules/world/npc_hunt.go:291 (sumPlayerInvParam, the non-ScriptState mirror)

All three now go through int(int32(iv)) which sign-extends correctly.

Pinned by TestOcParamInt_NegativeSignPreserved in pkg/script: param
encoded as uint32(0xFFFFFFFC) pops as int(-4). Pre-fix: 4294967292.

Routed in-scope-stretch per smoke_surfaces_adjacent_divergences (~30
LOC threshold met: 3 production lines + 1 fixture row + 1 ParamType
fixture entry + 1 test ≈ 25 LOC). Future audit should consider
fixing at the boundary (DecodeParams stores int32 instead of uint32),
but that would churn every test fixture using ParamMap{key: uint32(N)}
literals — out of scope here.

+++ modules/world/npc_hunt.go
@@ -288,7 +288,9 @@ func invTotalParam(inv *inventory.Inventory, param int,
 		}
 		if v, ok := ot.Params[uint32(param)]; ok {
 			if iv, ok := v.(uint32); ok {
-				total += int(iv)
+				// NAI-122 in-scope-stretch: sign-extend through int32.
+				// See paramLookup in pkg/script/handlers_config.go.
+				total += int(int32(iv))
 				continue
 			}
 		}
